requirements.txt
boto3>=1.35.0
faker>=25.0.0
python-dotenv>=1.0.1
